### 2025-10-29 #9323 @editorialbot
[PRE REVIEW]: stan-playground: Run Stan models directly in your browser.
**Submitting author:** @WardBrian (<a href="http://orcid.org/0000-0002-9841-3342">Brian Ward</a>)
**Repository:** https://github.com/flatironinstitute/stan-playground
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0 (continuous releases)
