### 2025-10-29 eae3177
Update citations for posterior, disease modeling case study
