+++ paper/paper.bib
@@ -104,13 +104,18 @@
     year = {2025},
     note = {R package version 1.6.1},
     url = {https://mc-stan.org/posterior/},
+    doi = {10.32614/cran.package.posterior}
 }
 
 @article{Grinsztajn2020,
-    title={Bayesian Workflow for disease transmission modeling in {Stan}},
-    volume={7},
-    url={https://mc-stan.org/learn-stan/case-studies/boarding_school_case_study.html},
-    journal={Stan Case Studies},
-    author={Grinsztajn, Léo and Semenova, Elizaveta and Margossian, Charles C. and Riou, Julien},
-    year={2020}
+    author = {Grinsztajn, Léo and Semenova, Elizaveta and Margossian, Charles C. and Riou, Julien},
+    title = {Bayesian workflow for disease transmission modeling in Stan},
+    journal = {Statistics in Medicine},
+    volume = {40},
+    number = {27},
+    pages = {6209-6234},
+    keywords = {Bayesian workflow, compartmental models, epidemiology, infectious diseases},
+    doi = {https://doi.org/10.1002/sim.9164},
+    url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9164},
+    year = {2021}
 }
