### 2025-10-29 #9323 @WardBrian
Besides the bots' suggestions, I think the following papers are potential sources of reviewers based on similar technology stacks:

- [textAnnotatoR: An Interactive Text Annotation Tool with 'shiny' GUI](https://github.com/openjournals/joss-reviews/issues/7971)
- [Trash AI: A Web GUI for Serverless Computer Vision Analysis of Images of Trash](https://github.com/openjournals/joss-reviews/issues/5136)
- [Powering single-cell analyses in the browser with WebAssembly](https://github.com/openjournals/joss-reviews/issues/5603)
- [ShUShER: private browser-based placement of sensitive genome samples on phylogenetic trees](https://github.com/openjournals/joss-reviews/issues/3677)

The following also serve similar communities and may provide familiarity with Stan:

- [The stantargets R package: a workflow framework for efficient reproducible Stan-powered Bayesian data analysis pipelines](https://github.com/openjournals/joss-reviews/issues/3193)
- [fishStan: Hierarchical Bayesian models for fisheries](https://github.com/openjournals/joss-reviews/issues/3444)
- [**geostan**: An R package for Bayesian spatial analysis](https://github.com/openjournals/joss-reviews/issues/4716)
