+++ paper/paper.bib
@@ -115,7 +115,7 @@
     number = {27},
     pages = {6209-6234},
     keywords = {Bayesian workflow, compartmental models, epidemiology, infectious diseases},
-    doi = {https://doi.org/10.1002/sim.9164},
+    doi = {10.1002/sim.9164},
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9164},
     year = {2021}
 }
