### 2025-11-03 87ac867
Use more digits in data.R JSON production
