### 2025-11-03 f3c9e4b
Print more digits in data.R's json conversion
