+++ gui/src/app/core/Scripting/webR/data_postamble.R
@@ -2,6 +2,6 @@
 if (typeof(data) != "list") {
     stop("[stan-playground] data must be a list")
 }
-.SP_DATA <- jsonlite::toJSON(data, pretty = TRUE, auto_unbox = TRUE, digits = NA)
+.SP_DATA <- jsonlite::toJSON(data, pretty = TRUE, auto_unbox = TRUE, digits = 18)
 # invisible prevents printing of the output
 invisible(.SP_DATA)
