### 2025-11-03 b75e06f
Move brms code to avoid accidental installs
