### 2025-11-03 cd00ce7
Remove unnecessary waits
