### 2025-11-28 #9323 @jflournoy
Yes, I can.

Sent from [Proton Mail](https://proton.me/mail/home) for Android.
