### 2025-11-30 #9531 @editorialbot
[REVIEW]: stan-playground: Run Stan models directly in your browser
**Submitting author:** @WardBrian (<a href="http://orcid.org/0000-0002-9841-3342">Brian Ward</a>)
**Repository:** https://github.com/flatironinstitute/stan-playground
**Branch with paper.md** (empty if default branch): joss
**Version:** v2026.02.10
