### 2025-11-30 #9531 @nkrusch
Hi @WardBrian, there is a dot at the end of the paper title. I suggest you remove this dot (edit the [paper.md](https://github.com/flatironinstitute/stan-playground/blob/46b4e35d1a0daf5dd8df0f2e27b2d44263f45063/paper/paper.md?plain=1#L2)).
