### 2025-12-02 1912398
Consistent capitalization of GitHub
