### 2025-12-02 d838703
Add CONTRIBUTING.md
