### 2025-12-02 90934c8
Apply suggestions from code review
Co-authored-by: Jeff Soules <jsoules@flatironinstitute.org>
