### 2026-01-12 #9531 @jflournoy
This is an exceptional contribution that broadens access to Stan and allows diverse educational and practical uses. The authors are thoughtful in explaining the benefits of this mode of working and provide multiple flexible ways of using the software. The amount of information and capabilities available via the web interface is really impressive. Setting up local compilation via the Docker container worked flawlessly.

- [ ] With regards to installation, I would suggest that the authors provide instructions for how to install the GUI. In fact, it seems to me that this front-end is precisely the software being advertised: the browser interface which allows the sampling to be done. This software is included in the repository but without clear reference or instructions.
- [ ]  Similarly, the authors should spend a few more sentences helping users understand the dependencies for local installation via Docker, as well as browser dependencies and compatibility for the front-end.
- [ ] Tests for the GUI seem to be present but there are no instructions for running them. There do not seem to be tests for the backend.
- [ ] I do not see any direct discussion of the state of the field with regard to similar software. It wouldn't surprise me if this is really in a class of its own but it would be good to hear what the authors have to say about their understanding of similar efforts.

The above comments are almost all about the presentation of this extremely useful and impressive work. I think they would make the work more accessible. Kudos to the authors.
