### 2026-01-13 85e6d8b
Once-over of installation documentation
