+++ README.md
@@ -33,6 +33,22 @@ how one can build the docker image from scratch.
 The website frontend is a Vitejs/React project located in the `gui/` folder.
 The [README](./gui/README.md) there gives information on how to get started running a local development server.
 
+## Installation
+
+No installation is required, simply navigate to https://stan-playground.flatironinstitute.org using
+a recent version of your favorite browser (tested in Chrome, Firefox, and Safari -- please let us know if you
+find a browser that does not work!).
+
+That being said, it is of course possible to run both the backend and website locally
+and get the same set of features.
+
+If you'd like to run a local compilation server, consult [the documentation](https://flatironinstitute.github.io/stan-playground/compilation_server_setup.html).
+The only dependency needed is [Docker](https://www.docker.com/).
+
+If you'd like to run a local copy of the website, consult the
+[instructions in the `gui/` folder](./gui/README.md). The only dependency
+needed to get started is [Yarn](https://yarnpkg.com/).
+
 ## License
 
 Apache License 2.0
