### 2026-01-13 bbd9600
Apply suggestions from code review
Co-authored-by: Jeff Soules <jsoules@flatironinstitute.org>
