### 2026-01-13 2a99516
Add section on prior work/similar projects
