+++ paper/paper.bib
@@ -119,3 +119,11 @@
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9164},
     year = {2021}
 }
+
+@Manual{Gabry2025,
+  title = {shinystan: Interactive Visual and Numerical Diagnostics and Posterior Analysis for Bayesian Models},
+  author = {Jonah Gabry and Duco Veen},
+  year = {2025},
+  note = {R package version 2.7.0},
+  url = {https://mc-stan.org/shinystan/},
+}
