### 2026-01-13 flatironinstitute/stan-playground#339 @WardBrian
Tests for backend server
These were requested in the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9531).

I think they would be nice-to-have, but the main benefit would be if we anticipated making changes to the way the backend worked, which thus far we have mostly avoided outside of version bumps.

Some manual testing will always be important to verify that the actual returned webassembly blobs work as expected, but we could unit test the endpoint API surface.
