### 2026-01-14 a41a42a
Update to Stan 2.38
