### 2026-01-22 #9531 @jflournoy
@WardBrian

> Do you see a use case where someone would want to not use the public website? The instructions we give [for developers](https://github.com/flatironinstitute/stan-playground/tree/main/gui#local-development) could be adapted to this, but I want to make sure I understand the intended audience you are foreseeing.

Yeah, I think the developer instructions are appropriate. I'm primarily thinking that a submission to a journal of open source software should include instructions for running the software, for whatever reason. I can imagine wanting to run my own GUI for reasons of privacy and the like. Who knows.

> If we expand our scope a bit, there is a constellation of websites that we pulled inspiration from:
> godbolt.org for C/C++ exporation, jsfiddle for interactive web development, and the Rust Playground (which provided some inspiration for the name as well as the functionality). These all share an ethos with Stan Playground, though the technical contributions differ (godbolt.org, for example, performs almost all computations on external servers)

I think this is fine. Just to help the reader understand where this software falls in a broader ecosystem. Thanks for thinking through that.

Thank you for engaging with the review!
