### 2026-01-23 6d99f45
Add sentence on local running
