### 2026-01-23 #9531 @WardBrian
@jflournoy I've added a sentence to mention the possibility of running the gui locally, but I've left the details in the (recently improved) repository READMEs,

Please take some time to re-review when you are able!
