+++ gui/src/app/util/files.ts
@@ -1,0 +1,8 @@
+export type File = { name: string; content: Uint8Array };
+
+export const encodeTextFile = (name: string, content: string): File => {
+  return {
+    name,
+    content: new TextEncoder().encode(content + "\n"),
+  };
+};
