### 2026-02-09 7eacf8f
Wire up extra data files to data model, serialization
