+++ gui/src/app/util/serializeAsZip.ts
@@ -2,7 +2,7 @@ import JSZip from "jszip";
 
 export const serializeAsZip = async (
   folderName: string,
-  files: { [key: string]: string } = {},
+  files: { [key: string]: string | Uint8Array } = {},
 ): Promise<[Blob, string]> => {
   const zip = new JSZip();
   const folder = zip.folder(folderName);
@@ -11,7 +11,10 @@ export const serializeAsZip = async (
   }
 
   Object.entries(files).forEach(([name, content]) => {
-    if (content.trim() !== "") {
+    if (typeof content === "string") {
+      content = content.trim();
+    }
+    if (content.length > 0) {
       folder.file(name, content);
     }
   });
