### 2026-02-09 376436d
Clean up zip creation
