+++ gui/src/app/util/serializeAsZip.ts
@@ -3,7 +3,7 @@ import JSZip from "jszip";
 export const serializeAsZip = async (
   folderName: string,
   files: { [key: string]: string | Uint8Array } = {},
-): Promise<[Blob, string]> => {
+): Promise<Blob> => {
   const zip = new JSZip();
   const folder = zip.folder(folderName);
   if (!folder) {
@@ -20,5 +20,5 @@ export const serializeAsZip = async (
   });
   const zipBlob = await zip.generateAsync({ type: "blob" });
 
-  return [zipBlob, folderName];
+  return zipBlob;
 };
