### 2026-02-09 #9531 @nkrusch
@WardBrian please fix this missing DOI in the bib

> ❌ MISSING DOIs
>
> - 10.32614/cran.package.shinystan may be a valid DOI for title: shinystan: Interactive Visual and Numerical Diagno...
