### 2026-02-10 #9531 @WardBrian
Hi @nkrusch --

I have merged your PR for the citation fixes and followed the rest of those steps.

The tagged release can be found at
https://github.com/flatironinstitute/stan-playground/releases/tag/v2026.02.10

The archive can be found at
https://doi.org/10.5281/zenodo.18598808
