### 2026-02-10 aca2d99
More code re-use, tests
