+++ gui/src/app/util/files.ts
@@ -1,24 +1,41 @@
 export type File = { name: string; content: Uint8Array };
 
+const encoder = new TextEncoder();
+
 export const encodeTextFile = (name: string, content: string): File => {
   return {
     name,
-    content: new TextEncoder().encode(content + "\n"),
+    content: encoder.encode(content + "\n"),
   };
 };
 
-export const base64encode = (f: File): { name: string; content: string } => {
+const decoder = new TextDecoder();
+
+export const tryDecodeText = (content: Uint8Array): string | undefined => {
+  try {
+    return decoder.decode(content);
+  } catch {
+    return undefined;
+  }
+};
+
+export const base64encode = (
+  f: File,
+): { name: string; b64contents: string } => {
   return {
     name: f.name,
-    content: btoa(
+    b64contents: btoa(
       f.content.reduce((data, byte) => data + String.fromCharCode(byte), ""),
     ),
   };
 };
 
-export const base64decode = (d: { name: string; content: string }): File => {
+export const base64decode = (d: {
+  name: string;
+  b64contents: string;
+}): File => {
   return {
     name: d.name,
-    content: Uint8Array.from(atob(d.content), (c) => c.charCodeAt(0)),
+    content: Uint8Array.from(atob(d.b64contents), (c) => c.charCodeAt(0)),
   };
 };
