### 2026-02-11 657217c
Avoid serializing file writes in webr/pyodide
