### 2026-02-11 271cf88
Update eslint config, check in CI
