### 2026-02-11 a5ae39a
Validate uploaded files to avoid smashing
