### 2026-02-12 6aa57e1
Hook lints and cleanup
