+++ gui/eslint.config.js
@@ -1,7 +1,10 @@
 import js from "@eslint/js";
 import globals from "globals";
+
 import tseslint from "typescript-eslint";
 import pluginReact from "eslint-plugin-react";
+import reactHooks from "eslint-plugin-react-hooks";
+
 import { defineConfig } from "eslint/config";
 
 export default defineConfig([
@@ -13,9 +16,18 @@ export default defineConfig([
   },
   tseslint.configs.recommended,
   pluginReact.configs.flat.recommended,
+  pluginReact.configs.flat["jsx-runtime"],
+  reactHooks.configs.flat.recommended,
+  {
+    settings: {
+      react: {
+        version: "19",
+      },
+    },
+  },
   {
     rules: {
-      "react/react-in-jsx-scope": "off",
+      "react-hooks/set-state-in-effect": "off",
       "@typescript-eslint/no-explicit-any": "off",
       "@typescript-eslint/no-unused-vars": [
         "error",
