### 2026-02-18 c972c30
Allow loading from a compressed ?project= query
