### 2026-02-19 7339414
Encapsulate
