### 2026-02-20 454a24f
Clean up parameter parsing logic for projects
