+++ gui/src/app/core/Project/ProjectSerialization.ts
@@ -164,27 +164,21 @@ export const deserializeZipToFiles = async (zipBuffer: Uint8Array) => {
 const loadMetaFromString = (
   data: ProjectDataModel,
   json: string,
-  clearExisting: boolean = false,
 ): ProjectDataModel => {
   const newMeta = JSON.parse(json);
   if (!isProjectMetaData(newMeta)) {
     throw new Error("Deserialized meta is not valid");
   }
-  const newMetaMember = clearExisting
-    ? { ...newMeta }
-    : { ...data.meta, ...newMeta };
+  const newMetaMember = { ...data.meta, ...newMeta };
   return { ...data, meta: newMetaMember };
 };
 
 const loadSamplingOptsFromString = (
   data: ProjectDataModel,
   json: string,
-  clearExisting: boolean = false,
 ): ProjectDataModel => {
   const newSampling = parseSamplingOpts(json);
-  const newSamplingOptsMember = clearExisting
-    ? { ...newSampling }
-    : { ...data.samplingOpts, ...newSampling };
+  const newSamplingOptsMember = { ...data.samplingOpts, ...newSampling };
   return { ...data, samplingOpts: newSamplingOptsMember };
 };
 
@@ -220,11 +214,11 @@ const loadExtraDataFilesFromString = (
 };
 
 export const loadFromProjectFiles = (
-  data: ProjectDataModel,
   files: Partial<FieldsContentsMap>,
-  clearExisting: boolean = false,
+  previous: ProjectDataModel | undefined = undefined,
 ): ProjectDataModel => {
-  let newData = clearExisting ? initialDataModel : data;
+  let newData = previous ? previous : structuredClone(initialDataModel);
+
   if (Object.keys(files).includes("meta")) {
     newData = loadMetaFromString(newData, files.meta ?? "");
     delete files["meta"];
