### 2026-02-23 475db10
Break up fetchRemoteProject
