+++ gui/src/app/core/Project/ProjectQueryLoading.ts
@@ -2,6 +2,8 @@ import { mapFileContentsToModel } from "@SpCore/Project/FileMapping";
 import loadFilesFromGist from "@SpUtil/gists/loadFilesFromGist";
 import {
   ProjectDataModel,
+  ProjectKnownFiles,
+  SamplingOpts,
   defaultSamplingOpts,
   initialDataModel,
   parseSamplingOpts,
@@ -64,150 +66,135 @@ export const fromQueryParams = (searchParams: URLSearchParams) => {
   return queries;
 };
 
+const queryParamToDataModelFieldMap = {
+  [QueryParamKeys.StanFile]: ProjectKnownFiles.STANFILE,
+  [QueryParamKeys.DataFile]: ProjectKnownFiles.DATAFILE,
+  [QueryParamKeys.AnalysisPyFile]: ProjectKnownFiles.ANALYSISPYFILE,
+  [QueryParamKeys.AnalysisRFile]: ProjectKnownFiles.ANALYSISRFILE,
+  [QueryParamKeys.DataPyFile]: ProjectKnownFiles.DATAPYFILE,
+  [QueryParamKeys.DataRFile]: ProjectKnownFiles.DATARFILE,
+} as const;
+
 export const queryStringHasParameters = (query: QueryParams) => {
   return Object.values(query).some((v) => v !== null);
 };
 
 export const fetchRemoteProject = async (query: QueryParams) => {
-  const projectParam = query.project;
-
-  if (projectParam) {
-    if (projectParam.startsWith("https://gist.github.com/")) {
-      try {
-        const contentLoadedFromGist = await loadFilesFromGist(projectParam);
-        const dataFromGist = loadFromProjectFiles(
-          mapFileContentsToModel(contentLoadedFromGist.files),
-        );
-        dataFromGist.meta.title = contentLoadedFromGist.description;
-        return persistStateToEphemera(dataFromGist);
-      } catch (err) {
-        console.error("Failed to load content from gist", err);
-        alert(`Failed to load content from gist ${projectParam}`);
-      }
-    } else if (hasKnownProjectParameterPrefix(projectParam)) {
-      try {
-        const dataFromParam = deserializeProjectFromURLParameter(projectParam);
-        if (dataFromParam) {
-          return persistStateToEphemera(dataFromParam);
-        } else {
-          throw new Error("Failed to deserialize project from URL parameter");
-        }
-      } catch (err) {
-        console.error("Failed to load content from project string", err);
-        alert("Failed to load content from compressed project");
-      }
-    } else {
-      console.error("Unsupported project parameter type", projectParam);
-    }
+  if (query.project) {
     // other parameters are ignored whenever project= is set
-    return initialDataModel;
+    return await loadFromProjectParameter(query.project);
   }
 
   const data: ProjectDataModel = structuredClone(initialDataModel);
-  const stanFilePromise = query.stan
-    ? tryFetch(query.stan)
-    : Promise.resolve(data.stanFileContent);
-  const dataFilePromise = query.data
-    ? tryFetch(query.data)
-    : Promise.resolve(data.dataFileContent);
-  const analysisPyFilePromise = query["analysis_py"]
-    ? tryFetch(query["analysis_py"])
-    : Promise.resolve(data.analysisPyFileContent);
-  const analysisRFilePromise = query["analysis_r"]
-    ? tryFetch(query["analysis_r"])
-    : Promise.resolve(data.analysisRFileContent);
-  const dataPyFilePromise = query["data_py"]
-    ? tryFetch(query["data_py"])
-    : Promise.resolve(data.dataPyFileContent);
-  const dataRFilePromise = query["data_r"]
-    ? tryFetch(query["data_r"])
[…]
