### 2026-02-25 #9531 @crvernon
Very nice paper and product.  Thank you for your attention to detail.
